Manage Sparkle through Swift Package Manager (SPM)
Installed version: 1.26.0

Note: this will not really work for the moment, as I have no access to
the keypair needed for signing an update, nor control the website which
publishes the appcast. My intent is to generate in a near future a new
pair of EdDSA keys and switch to Sparkle 2.